other: v2 actor 基本模型

+++ toolkit/mask/dynamic_mask_benchmark_test.go
@@ -10,7 +10,7 @@ func BenchmarkDynamicMask_Set(b *testing.B) {
 	b.Run("Set", func(b *testing.B) {
 		var m mask.DynamicMask
 		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
+		for i := uint64(0); i < uint64(b.N); i++ {
 			m.Set(i)
 		}
 	})
@@ -19,7 +19,7 @@ func BenchmarkDynamicMask_Set(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			var m mask.DynamicMask
-			for j := 0; j < 64; j++ {
+			for j := uint64(0); j < 64; j++ {
 				m.Set(j)
 			}
 		}
@@ -29,7 +29,7 @@ func BenchmarkDynamicMask_Set(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			var m mask.DynamicMask
-			for j := 0; j < 128; j++ {
+			for j := uint64(0); j < 128; j++ {
 				m.Set(j)
 			}
 		}
@@ -39,7 +39,7 @@ func BenchmarkDynamicMask_Set(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			var m mask.DynamicMask
-			for j := 0; j < 256; j++ {
+			for j := uint64(0); j < 256; j++ {
 				m.Set(j)
 			}
 		}
@@ -50,7 +50,7 @@ func BenchmarkDynamicMask_Set(b *testing.B) {
 func BenchmarkDynamicMask_Bits(b *testing.B) {
 	b.Run("Bits", func(b *testing.B) {
 		var m mask.DynamicMask
-		for i := 0; i < 1000; i++ {
+		for i := uint64(0); i < 1000; i++ {
 			m.Set(i)
 		}
 		b.ResetTimer()
@@ -64,7 +64,7 @@ func BenchmarkDynamicMask_Bits(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			var m mask.DynamicMask
-			for j := 0; j < 64; j++ {
+			for j := uint64(0); j < 64; j++ {
 				m.Set(j)
 			}
 			m.Bits()
@@ -75,12 +75,12 @@ func BenchmarkDynamicMask_Bits(b *testing.B) {
 func BenchmarkDynamicMask_Has(b *testing.B) {
 	b.Run("Has", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
 		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
+		for i := uint64(0); i < uint64(b.N); i++ {
 			m.Has(i)
 		}
 	})
@@ -89,13 +89,13 @@ func BenchmarkDynamicMask_Has(b *testing.B) {
 func BenchmarkDynamicMask_HasAny(b *testing.B) {
 	b.Run("HasAny", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			m.HasAny([]int{1, 2, 3})
+			m.HasAny([]uint64{1, 2, 3})
 		}
 	})
 }
@@ -103,7 +103,7 @@ func BenchmarkDynamicMask_HasAny(b *testing.B) {
 func BenchmarkDynamicMask_Equal(b *testing.B) {
 	b.Run("Equal", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
@@ -117,13 +117,13 @@ func BenchmarkDynamicMask_Equal(b *testing.B) {
 func BenchmarkDynamicMask_EqualBits(b *testing.B) {
 	b.Run("EqualBits", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 3; j++ {
+		for j := uint64(0); j < 3; j++ {
 			m.Set(j)
 		}
 
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			m.EqualBits([]int{1, 2, 3})
+			m.EqualBits([]uint64{1, 2, 3})
 		}
 	})
 }
@@ -131,7 +131,7 @@ func BenchmarkDynamicMask_EqualBits(b *testing.B) {
 func BenchmarkDynamicMask_Clear(b *testing.B) {
 	b.Run("Clear", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
@@ -145,7 +145,7 @@ func BenchmarkDynamicMask_Clear(b *testing.B) {
 func BenchmarkDynamicMask_Clone(b *testing.B) {
 	b.Run("Clone", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
@@ -159,12 +159,12 @@ func BenchmarkDynamicMask_Clone(b *testing.B) {
 func BenchmarkDynamicMask_Del(b *testing.B) {
 	b.Run("Del", func(b *testing.B) {
 		var m mask.DynamicMask
-		for j := 0; j < 64; j++ {
+		for j := uint64(0); j < 64; j++ {
 			m.Set(j)
 		}
 
 		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
+		for i := uint64(0); i < uint64(b.N); i++ {
 			m.Del(i)
 		}
 	})
@@ -174,7 +174,7 @@ func BenchmarkDynamicMask_SetAndDel(b *testing.B) {
 	b.Run("SetAndDel", func(b *testing.B) {
 		var m mask.DynamicMask
 		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
+		for i := uint64(0); i < uint64(b.N); i++ {
 			m.Set(i)
 			m.Del(i)
 		}
